// src/base/types/form-ad.ts
export interface FormData {
  title: string;
  description: string;
  category: string;
  subcategory: string;
  price: string;
  currency: string;
  negotiable: boolean;
  condition: string;
  department: string;
  province: string;
  district: string;
  contactName: string;
  contactPhone: string;
  contactEmail: string;
  images: File[];
}